reorganized project structure and migration
