+++ src/app.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { typeOrmConfig } from './configs/typeorm.config';
-import { OffersModule } from './offers/offers.module';
-import { CoursesModule } from './course/course.module';
-import { CampusModule } from './campuses/campus.module';
-import { UniversitiesModule } from './university/university.module';
+import { OffersModule } from './modules/offers.module';
+import { CoursesModule } from './modules/course.module';
+import { CampusModule } from './modules/campus.module';
+import { UniversitiesModule } from './modules/university.module';
 
 @Module({
   imports: [
