+++ src/modules/offers.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { OfferRepostiroy } from './offers.repository';
-import { OffersService } from './offers.service';
+import { OfferRepostiroy } from '../repositories/offers.repository';
+import { OffersService } from '../services/offers.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([OfferRepostiroy])],
