+++ src/modules/university.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UniversityRepostiroy } from './university.repository';
-import { UniversityService } from './university.service';
+import { UniversityRepostiroy } from '../repositories/university.repository';
+import { UniversityService } from '../services/university.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UniversityRepostiroy])],
