+++ src/repositories/campus.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { Campus } from './campus.entity';
+import { Campus } from '../models/campus.entity';
 
 @EntityRepository(Campus)
 export class CampusRepository extends Repository<Campus> {}
