+++ src/repositories/offers.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { Offer } from './offers.entity';
+import { Offer } from '../models/offers.entity';
 
 @EntityRepository(Offer)
 export class OfferRepostiroy extends Repository<Offer> {}
