+++ src/repositories/university.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import { University } from './university.entity';
+import { University } from '../models/university.entity';
 
 @EntityRepository(University)
 export class UniversityRepostiroy extends Repository<University> {}
