+++ src/services/offers.service.ts
